Temporary add-in for powerups that change the next block the current player will receive next.
Added in BOMB and ANTIBOMB blocks when the current block has a darkened/lightened piece and clears a row with it.

Will eventually need the PowerUp interface and the corresponding classes for each of them.
replaceNextRad() (like bomb block, antibomb block)
beginAction() (unnecesary for bomb/antibomb)
endAction() (like blowup, blowdown)