Created PowerUp class to consolidate the various powerups and make it much easier to create new ones.
Currently implemented for Tetris and the TetrisHeuristicAI are:
BOMB
ANTIBOMB
SCRAMBLE (gives the "PowerUp" [more like debuff] to the opponent; it scrambles the bottom half of the opponent's board.)
REMOVE (which takes away a row from the bottom of your screen)

Tetrad currently controls how often/when the powerups appear, although that can and probably will be moved to PowerUp.

Switched the Tetris games back to JPanelBlockDisplay, although the BitmapBlockDisplay is pretty cool too. (Downside is that the new rows added at the bottom of the screen are always black now, so it's hard to see gaps.)